requirements.txt
sphinx   #==4.3.1
sphinx_rtd_theme==1.0.0
furo==2021.11.23
numpy==1.21.4
pandas==1.3.4
gensim==4.1.2
# spacy==3.2.1
plotly==5.4.0
umap-learn==0.5.2
loguru==0.5.3
pyarrow==6.0.1
dask==2021.11.2
hydra-core==1.1.1
spacy==3.4.4
#https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz